Move CustomAttributesMap to commons renamed to Properties
