+++ src/atom_space/AtomSpace.h
@@ -139,7 +139,7 @@ class AtomSpace {
      */
     char* add_node(const string& type,
                    const string& nam,
-                   const CustomAttributesMap& custom_attributes = {});
+                   const Properties& custom_attributes = {});
 
     /**
      * @brief Add a link to the AtomSpace.
@@ -154,7 +154,7 @@ class AtomSpace {
      */
     char* add_link(const string& type,
                    const vector<const Atom*>& targets,
-                   const CustomAttributesMap& custom_attributes = {});
+                   const Properties& custom_attributes = {});
 
     /**
      * @brief Commit changes to the AtomSpace.
