+++ src/atomdb/redis_mongodb/RedisMongoDB.cc
@@ -275,7 +275,7 @@ vector<string> RedisMongoDB::links_exist(const vector<string>& link_handles) {
 
 char* RedisMongoDB::add_node(const char* type,
                              const char* name,
-                             const atomdb_api_types::CustomAttributesMap& custom_attributes) {
+                             const Properties& custom_attributes) {
     // TODO: Implement add_node logic
     return NULL;
 }
@@ -283,7 +283,7 @@ char* RedisMongoDB::add_node(const char* type,
 char* RedisMongoDB::add_link(const char* type,
                              char** targets,
                              size_t targets_size,
-                             const atomdb_api_types::CustomAttributesMap& custom_attributes) {
+                             const Properties& custom_attributes) {
     // TODO: Implement add_link logic
     return NULL;
 }
