+++ src/atomdb/redis_mongodb/RedisMongoDB.h
@@ -57,11 +57,11 @@ class RedisMongoDB : public AtomDB {
     std::vector<std::string> links_exist(const std::vector<std::string>& link_handles);
     char* add_node(const char* type,
                    const char* name,
-                   const atomdb_api_types::CustomAttributesMap& custom_attributes = {});
+                   const Properties& custom_attributes = {});
     char* add_link(const char* type,
                    char** targets,
                    size_t targets_size,
-                   const atomdb_api_types::CustomAttributesMap& custom_attributes = {});
+                   const Properties& custom_attributes = {});
     vector<shared_ptr<atomdb_api_types::AtomDocument>> get_atom_documents(const vector<string>& handles,
                                                                           const vector<string>& fields);
